BlobTriggerFunction/requirements.txt
azure-functions
pandas
openpyxl
azure-storage-blob
